db_service/db_init.py
from typing import AsyncGenerator
from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
from config_reader import config
from db_service.models import Base


engine = create_async_engine(config.db_url.get_secret_value(), echo=True)

async_session_maker = async_sessionmaker(
    bind=engine, class_=AsyncSession, expire_on_commit=False
)


async def get_async_session() -> AsyncGenerator[AsyncSession, None]:
    async with async_session_maker() as session:
        yield session


async def create_all_tables():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.create_all)


async def delete_all_tables():
    async with engine.begin() as conn:
        await conn.run_sync(Base.metadata.drop_all)
